Moved `data_manager` to Datamanager module
I'm not sure why `data_manager` was in store._metadata.

I'm moving it now because I'm also removing the other code from
store._metadata, so that module can be deleted.
