+++ tests/dataobjects/test_datamanager.py
@@ -1,4 +1,4 @@
-from openghg.store import data_manager
+from openghg.dataobjects import data_manager
 from openghg.store.base import Datasource
 from openghg.objectstore.metastore import open_metastore
 from openghg.retrieve import search_surface
